Fix lable axis ticks on report graphing
