+++ luto/tools/report/create_html.py
@@ -4,7 +4,7 @@ import pandas as pd
 from glob import glob
 import argparse
 
-from tools.helper_func import add_data_2_html
+from tools.helper_func import add_data_2_html, add_settings_2_html
 
 
 # # setting up working directory to root dir
@@ -46,38 +46,8 @@ shutil.copytree('luto/tools/report/tools/template_html',
 #                Write data to HTML                #
 #################################################### 
 
-import os
-import pandas as pd
-
-from lxml import etree
-from  lxml.etree import Element
-
-# Get the tree of the index page
-index_tree = etree.parse(f"{REPORT_DIR}/REPORT_HTML/index.html", 
-                         etree.HTMLParser())
-
-# Get the txt of the running settings
-with open(f"{RAW_DATA_ROOT}/model_run_settings.txt",'r') as f:
-    settings_txt = f.read()
-    
-# Replace the inner text of the <pre id="settingsTxt"></pre> with 
-# the settings txt and make it to be invisible
-settings_pre = index_tree.xpath('//pre[@id="settingsTxt"]')[0]
-settings_pre.text = settings_txt
-settings_pre.attrib['style'] = 'display:none'
-  
-    
-# Write the index page
-index_tree.write(f"{REPORT_DIR}/REPORT_HTML/index.html", 
-                 pretty_print=True,
-                 encoding='utf-8',
-                 method='html')
-    
-
-    
-    
-
- 
+# Add settings to the home page
+add_settings_2_html(REPORT_DIR, RAW_DATA_ROOT)
 
 # Get all html files needs data insertion
 html_df = pd.DataFrame([['production',f"{REPORT_DIR}/REPORT_HTML/pages/production.html"],
