+++ luto/tools/report/tools/helper_func.py
@@ -367,6 +367,29 @@ def list_all_files(directory):
     return file_list
 
 
+def add_settings_2_html(REPORT_DIR:str, RAW_DATA_ROOT:str)->None:
+    # Get the tree of the index page
+    index_tree = etree.parse(f"{REPORT_DIR}/REPORT_HTML/index.html", 
+                            etree.HTMLParser())
+
+    # Get the txt of the running settings
+    with open(f"{RAW_DATA_ROOT}/model_run_settings.txt",'r') as f:
+        settings_txt = f.read()
+        
+    # Replace the inner text of the <pre id="settingsTxt"></pre> with 
+    # the settings txt and make it to be invisible
+    settings_pre = index_tree.xpath('//pre[@id="settingsTxt"]')[0]
+    settings_pre.text = settings_txt
+    settings_pre.attrib['style'] = 'display:none'
+    
+    
+    # Write the index page
+    index_tree.write(f"{REPORT_DIR}/REPORT_HTML/index.html", 
+                    pretty_print=True,
+                    encoding='utf-8',
+                    method='html')
+    
+
 def add_data_2_html(html_path:str, data_pathes:list)->None:
     """
     Adds data from multiple files to an HTML file.
