+++ luto/tools/report/tools/template_html/pages/economics.html
@@ -45,7 +45,7 @@
         <div class="fig" id="economics_1_revenue_1_Irrigation_wide"></div>
 
         <!-- Chart:economics_1_revenue_2_Source_wide -->
-        <div class="fig" id="economics_1_revenue_2_Source_wide"></div>
+        <!-- <div class="fig" id="economics_1_revenue_2_Source_wide"></div> -->
 
         <!-- Chart:economics_1_revenue_3_Source_type_wide -->
         <div class="fig" id="economics_1_revenue_3_Source_type_wide"></div>
