+++ luto/tools/report/tools/template_html/pages/water_usage.html
@@ -38,12 +38,6 @@
     <div class="content">
         <h1>Water Use</h1>
 
-        <!-- Chart:water_1_percent_to_limit -->
-        <div class="fig" id="water_1_percent_to_limit"></div>
-
-        <!-- Chart:water_2_volum_to_limit -->
-        <div class="fig" id="water_2_volum_to_limit"></div>
-
         <!-- Chart:water_3_volum_by_sector -->
         <div class="fig" id="water_3_volum_by_sector"></div>
 
@@ -52,6 +46,12 @@
 
         <!-- Chart:water_5_volum_by_irrigation -->
         <div class="fig" id="water_5_volum_by_irrigation"></div>
+
+        <!-- Chart:water_1_percent_to_limit -->
+        <div class="fig" id="water_1_percent_to_limit"></div>
+
+        <!-- Chart:water_2_volum_to_limit -->
+        <div class="fig" id="water_2_volum_to_limit"></div>
